requirements.txt
openai
langchain
pinecone-client
python-dotenv
tiktoken
matplotlib
pypdf
docx2txt
wikipedia